Look and Polish
Some refactoring done regarding the minigame buttons.

Added a free skybox from the asset store.

+++ FYP - Spoken Word/Assets/Scripts/Minigame/Goal.cs
@@ -6,5 +6,5 @@ public class Goal : MonoBehaviour
 {
 	public int pointsAwarded;
 	public int bonus;
-	public int bonusMultiplier;
+	public float bonusMultiplier;
 }
